requirements.txt
anthropic>=0.18.0
python-docx>=1.1.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.1.0
